LibWeb: Split ListStyleStyleValue out of StyleValue.{h,cpp}
